+++ R/autoplot-density.R
@@ -88,7 +88,7 @@ tidy_autoplot <- function(.data, .plot_type = "density", .line_size = .5,
     "tidy_zero_truncated_binomial", "tidy_zero_truncated_negative_binomial",
     "tidy_pareto_single_parameter", "tidy_pareto", "tidy_inverse_pareto",
     "tidy_generalized_pareto","tidy_paralogistic", "tidy_inverse_exponential",
-    "tidy_inverse_gamma"
+    "tidy_inverse_gamma","tidy_inverse_weibull"
   )) {
     rlang::abort("The data passed must come from a `tidy_` distribution function.")
   }
@@ -163,7 +163,8 @@ tidy_autoplot <- function(.data, .plot_type = "density", .line_size = .5,
       )
     } else if (atb$tibble_type %in% c(
         "tidy_paralogistic",
-        "tidy_inverse_gamma"
+        "tidy_inverse_gamma",
+        "tidy_inverse_weibull"
         )
       ){
       paste0("Shape: ", atb$.shape, " - ",