+++ R/random-tidy-weibull.R
@@ -1,6 +1,8 @@
 #' Tidy Randomly Generated Weibull Distribution Tibble
 #'
-#' @family Data Generator
+#' @family Continuous Distribution
+#' @family Weibull
+#' @family Inverse Distribution
 #'
 #' @author Steven P. Sanderson II, MPH
 #'